# server/docs/openapi.yml
components:
  schemas:
    api.AccountInfoResponse:
      properties:
        data:
          $ref: "#/components/schemas/model.AccountInfo"
      required:
        - data
      type: object
...
    api.ConfigureProviderRequest:
      properties:
        config:
          additionalProperties:
            type: string
          type: object
        enabled:
          type: boolean
      required:
        - enabled
      type: object
    api.CreateDownloadRequest:
      properties:
        destination:
          example: gdrive:movies
          type: string
        dir:
          example: /downloads
          type: string
        engine:
          enum:
            - native
            - aria2
          type: string
        filename:
          example: my_file.zip
          type: string
        files:
          items:
            $ref: "#/components/schemas/model.DownloadFile"
          type: array
          uniqueItems: false
        hash:
          type: string
        headers:
          additionalProperties:
            type: string
          type: object
        provider:
          type: string
        proxies:
          items:
            $ref: "#/components/schemas/api.Proxy"
          type: array
          uniqueItems: false
        removeLocal:
          type: boolean
        selectedFiles:
          example:
            - 1
            - 2
          items:
            type: integer
          type: array
          uniqueItems: false
        split:
          type: integer
        torrentData:
          description: Fields For magnets
          type: string
        url:
          example: http://example.com/file.zip
          type: string
      required:
        - url
      type: object
...
    api.DownloadListResponse:
      properties:
        data:
          items:
            $ref: "#/components/schemas/model.Download"
          type: array
          uniqueItems: false
        meta:
          $ref: "#/components/schemas/api.Meta"
      required:
        - data
      type: object
    api.DownloadResponse:
      properties:
        data:
          $ref: "#/components/schemas/model.Download"
      required:
        - data
      type: object
    api.ErrorResponse:
      properties:
        code:
          example: 500
          type: integer
        error:
          example: error message
          type: string
      required:
        - code
        - error
      type: object
...
    api.Meta:
      properties:
        limit:
          example: 10
          type: integer
        offset:
          example: 0
          type: integer
        total:
          example: 100
          type: integer
      required:
        - limit
        - offset
        - total
      type: object
...
    api.ProviderHosts:
      properties:
        hosts:
          items:
            type: string
          type: array
          uniqueItems: false
      required:
        - hosts
      type: object
    api.ProviderHostsResponse:
      properties:
        data:
          $ref: "#/components/schemas/api.ProviderHosts"
      required:
        - data
      type: object
    api.ProviderListResponse:
      properties:
        data:
          items:
            $ref: "#/components/schemas/model.ProviderSummary"
          type: array
          uniqueItems: false
      required:
        - data
      type: object
    api.ProviderSchema:
      properties:
        configSchema:
          items:
            $ref: "#/components/schemas/provider.ConfigField"
          type: array
          uniqueItems: false
        name:
          type: string
      required:
        - configSchema
        - name
      type: object
    api.ProviderSchemaResponse:
      properties:
        data:
          $ref: "#/components/schemas/api.ProviderSchema"
      required:
        - data
      type: object
    api.Proxy:
      properties:
        type:
          type: string
        url:
          type: string
      type: object
...
    api.ResolveResult:
      properties:
        provider:
          type: string
        result:
          $ref: "#/components/schemas/provider.ResolveResult"
        url:
          type: string
      required:
        - provider
        - result
        - url
      type: object
    api.ResolveResultResponse:
      properties:
        data:
          $ref: "#/components/schemas/api.ResolveResult"
      required:
        - data
      type: object
    api.ResolveURLRequest:
      properties:
        headers:
          additionalProperties:
            type: string
          type: object
        torrentBase64:
          type: string
        url:
          example: http://example.com/file.zip
          type: string
      required:
        - url
      type: object
...
    model.AccountInfo:
      properties:
        expiresAt:
          type: string
        isPremium:
          type: boolean
        username:
          type: string
      type: object
...
    model.Download:
      properties:
        completedAt:
          type: string
        createdAt:
          type: string
        destination:
          type: string
        dir:
          type: string
        downloaded:
          example: 5242880
          type: integer
        engine:
          enum:
            - aria2
            - native
          type: string
        error:
          type: string
        eta:
          example: 10
          type: integer
        filename:
          type: string
        files:
          items:
            $ref: "#/components/schemas/model.DownloadFile"
          type: array
          uniqueItems: false
        headers:
          additionalProperties:
            type: string
          type: object
        id:
          example: d_a1b2c3d4
          type: string
        isMagnet:
          type: boolean
        magnetHash:
          type: string
        peerDetails:
          items:
            $ref: "#/components/schemas/model.Peer"
          type: array
          uniqueItems: false
        peers:
          type: integer
        provider:
          type: string
        proxies:
          items:
            $ref: "#/components/schemas/model.Proxy"
          type: array
          uniqueItems: false
        removeLocal:
          type: boolean
        resolvedUrl:
          type: string
        seeders:
          type: integer
        size:
          example: 10485760
          type: integer
        speed:
          example: 1024000
          type: integer
        split:
          description: Not Saved in DB
          type: integer
        startedAt:
          type: string
        status:
          $ref: "#/components/schemas/model.DownloadStatus"
        torrentData:
          type: string
        updatedAt:
          type: string
        uploadProgress:
          example: 50
          type: integer
        uploadSpeed:
          example: 512000
          type: integer
        uploadStatus:
          enum:
            - idle
            - running
            - complete
            - error
          type: string
        url:
          example: http://example.com/file.zip
          type: string
      required:
        - dir
        - downloaded
        - eta
        - filename
        - id
        - size
        - speed
        - status
        - url
      type: object
    model.DownloadFile:
      properties:
        createdAt:
          type: string
        downloaded:
          type: integer
        error:
          type: string
        id:
          type: string
        name:
          type: string
        path:
          description: 'relative path: "ubuntu/iso/file.iso"'
          type: string
        progress:
          type: integer
        size:
          type: integer
        status:
          $ref: "#/components/schemas/model.DownloadStatus"
        updatedAt:
          type: string
      type: object
...
    model.DownloadStatus:
      description: active, waiting, paused, uploading, complete, error
      enum:
        - active
        - waiting
        - paused
        - uploading
        - complete
        - error
        - allocating
        - processing
        - resolving
        - active
        - waiting
        - paused
        - uploading
        - complete
        - error
      type: string
      x-enum-varnames:
        - StatusActive
        - StatusWaiting
        - StatusPaused
        - StatusUploading
        - StatusComplete
        - StatusError
        - StatusAllocating
        - StatusProcessing
        - StatusResolving
...
    model.Peer:
      properties:
        downloadSpeed:
          type: integer
        ip:
          type: string
        isSeeder:
          type: boolean
        port:
          type: string
        uploadSpeed:
          type: integer
      required:
        - downloadSpeed
        - ip
        - isSeeder
        - port
        - uploadSpeed
      type: object
...
    model.ProviderSummary:
      properties:
        account:
          $ref: "#/components/schemas/model.AccountInfo"
        configured:
          type: boolean
        displayName:
          type: string
        enabled:
          type: boolean
        name:
          type: string
        priority:
          type: integer
        type:
          $ref: "#/components/schemas/model.ProviderType"
      type: object
    model.ProviderType:
      enum:
        - direct
        - debrid
        - filehost
      type: string
      x-enum-varnames:
        - ProviderTypeDirect
        - ProviderTypeDebrid
        - ProviderTypeFileHost
    model.Proxy:
      properties:
        type:
          enum:
            - all
            - downloads
            - uploads
            - magnets
          type: string
        url:
          type: string
      type: object
...
    provider.ConfigField:
      properties:
        description:
          type: string
        key:
          type: string
        label:
          type: string
        required:
          type: boolean
        type:
          description: text, password, select
          type: string
      type: object
    provider.ResolveResult:
      properties:
        error:
          type: string
        files:
          items:
            $ref: "#/components/schemas/model.DownloadFile"
          type: array
          uniqueItems: false
        hash:
          type: string
        headers:
          additionalProperties:
            type: string
          type: object
        isMagnet:
          type: boolean
        name:
          type: string
        size:
          type: integer
        url:
          type: string
      type: object
...
info:
  description: API for Gravity Download Manager
  title: Gravity API
  version: "1.0"
openapi: 3.1.0
paths:
  /downloads:
    get:
      description: Get a paginated list of downloads with optional status filtering
      parameters:
        - description: Comma-separated statuses to filter by
          in: query
          name: status
          schema:
            type: string
        - description: Max number of items to return
          in: query
          name: limit
          schema:
            type: integer
        - description: Offset for pagination
          in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.DownloadListResponse"
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: List downloads
      tags:
        - downloads
    post:
      description: Start a new download from a URL
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                - $ref: "#/components/schemas/api.CreateDownloadRequest"
                  description: Download request
                  summary: request
        description: Download request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.DownloadResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: Create download
      tags:
        - downloads
  /downloads/{id}:
    delete:
      description: Stop and delete a download, optionally removing files
      parameters:
        - description: Download ID
          in: path
          name: id
          required: true
          schema:
            type: string
        - description: If true, deletes files from disk
          in: query
          name: delete_files
          schema:
            type: boolean
      responses:
        "204":
          description: No Content
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: Delete download
      tags:
        - downloads
    get:
      description: Get details of a specific download by ID
      parameters:
        - description: Download ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.DownloadResponse"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Not Found
      summary: Get download
      tags:
        - downloads
  /downloads/{id}/pause:
    post:
      description: Pause an active download
      parameters:
        - description: Download ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: Pause download
      tags:
        - downloads
  /downloads/{id}/resume:
    post:
      description: Resume a paused download
      parameters:
        - description: Download ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: Resume download
      tags:
        - downloads
  /downloads/{id}/retry:
    post:
      description: Retry a failed or completed download
      parameters:
        - description: Download ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: Retry download
      tags:
        - downloads
...
  /providers:
    get:
      description: Get a list of all configured download providers
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ProviderListResponse"
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: List providers
      tags:
        - providers
  /providers/{name}:
    delete:
      description: Delete configuration for a specific provider
      parameters:
        - description: Provider name
          in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: Remove provider configuration
      tags:
        - providers
    get:
      description: Get the configuration schema for a specific provider
      parameters:
        - description: Provider name
          in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ProviderSchemaResponse"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Not Found
      summary: Get provider config schema
      tags:
        - providers
    put:
      description: Enable/disable or update configuration settings for a provider
      parameters:
        - description: Provider name
          in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                - $ref: "#/components/schemas/api.ConfigureProviderRequest"
                  description: Configuration request
                  summary: request
        description: Configuration request
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: Update provider configuration
      tags:
        - providers
  /providers/{name}/hosts:
    get:
      description: Get list of file hosts supported by a debrid provider
      parameters:
        - description: Provider name
          in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ProviderHostsResponse"
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: Get supported hosts
      tags:
        - providers
  /providers/{name}/status:
    get:
      description: Get current status and account info for a provider
      parameters:
        - description: Provider name
          in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.AccountInfoResponse"
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Internal Server Error
      summary: Get provider status
      tags:
        - providers
  /providers/resolve:
    post:
      description: Resolve a download URL through providers to get direct links
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                - $ref: "#/components/schemas/api.ResolveURLRequest"
                  description: Resolve request
                  summary: request
        description: Resolve request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ResolveResultResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.ErrorResponse"
          description: Not Found
      summary: Resolve URL
      tags:
        - providers
...
servers:
  - url: localhost:8080/api/v1